refactor(api): split endpoint feedback controller
