+++ apps/api/src/controllers/v2/feedback/daily-cap.ts
@@ -1,0 +1,6 @@
+import { config } from "../../../config";
+import { FeedbackRecordOptions } from "./internal-types";
+
+export function dailyCapFor(options: FeedbackRecordOptions): number {
+  return options.dailyCapCredits ?? config.FEEDBACK_DAILY_CAP_CREDITS;
+}
